+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased
 
+* [#916](https://github.com/Shopify/shopify_api/pull/916) Use new Metafield types API
+
 ## Version 9.5.1
 
 - [#891](https://github.com/Shopify/shopify_api/pull/891) Removed the upper bound on the `activeresource` dependency to allow apps to use the latest version